--- msbuild/Xamarin.Mac.Tasks/Tasks/Archive.cs
@@ -1,6 +1,0 @@
-﻿namespace Xamarin.Mac.Tasks
-{
-	public class Archive : ArchiveTaskBase
-	{
-	}
-}
